Allows the removal of default apps from the dock
No more photobooth, no more garageband sitting in the dock next to
RubyMine. This can be turned on/off via the configuration setting:
'dock_preferences'->'clear_apps'

Also allows for the dock size to be adjusted and makes orientation
adjustment optional via configuration setting

+++ sprout-osx-settings/recipes/dock_preferences.rb
@@ -1,15 +1,32 @@
-osx_defaults "set dock to be on left" do
-  domain "com.apple.dock"
-  key "orientation"
-  string node["dock_preferences"]["orientation"]
+osx_defaults "set dock to be on #{node['dock_preferences']['orientation']}" do
+  domain 'com.apple.dock'
+  key 'orientation'
+  string node['dock_preferences']['orientation']
+  only_if { node['dock_preferences']['orientation'] }
 end
 
-osx_defaults "set dock to autohide" do
-  domain "com.apple.dock"
-  key "autohide"
-  boolean node["dock_preferences"]["autohide"] && node["dock_preferences"]["autohide"] != 'false'
+osx_defaults 'set dock to autohide' do
+  domain 'com.apple.dock'
+  key 'autohide'
+  boolean node['dock_preferences']['autohide']
+  only_if { node['dock_preferences'].keys.include?('autohide') }
 end
 
-execute "relaunch dock" do
-  command "killall Dock"
+osx_defaults 'remove persistent apps from the dock' do
+  domain 'com.apple.dock'
+  key 'persistent-apps'
+  array []
+  only_if { node['dock_preferences']['clear_apps'] }
+end
+
+osx_defaults "adjusts dock size to #{node['dock_preferences']['tilesize']}" do
+  domain 'com.apple.dock'
+  key 'tilesize'
+  integer node['dock_preferences']['tilesize']
+  only_if { node['dock_preferences']['tilesize'] }
+end
+
+execute 'relaunch dock' do
+  command 'killall Dock'
+  only_if { !node['dock_preferences'].empty? }
 end
